change file composition and file name
